S3V60 - Complete
In this video we learned about the initState() method in StatefulWidgets which is a run-once function that allows us to "set up" a stateful widget. We used this function to set the value of our activeScreen property in the QuizState class. When this was done, we were able to tap the "Start Quiz" button and have the view change.