S5V101. Efficiently Rendering Long Lists with ListView
Introduction to the ListView widget by way of it's builder() function. The two properties are the itemCount Property which should be the length of the list it's drawing and the itemBuilder which defines the widget you actually want to return in your list as a function. Notice the anon function requires the context and an index to complete.